django_project/requirements.txt
boto3==1.17.108
botocore==1.20.108
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.17.1
gunicorn==20.1.0
idna==2.7
jmespath==0.10.0
Pillow==8.3.1
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.6
whitenoise==5.2.0